requirements.txt
blinker==1.6.2
click==8.1.3
colorama==0.4.6
Flask==2.3.2
importlib-metadata==6.6.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
numpy==1.24.3
pandas==2.0.2
pyodbc==4.0.39
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
pywin32==306
six==1.16.0
tzdata==2023.3
uuid==1.30
Werkzeug==2.3.4
zipp==3.15.0
